refactor(web): decompose studio screens
